requirements-no-pin.txt
hydra-core
loguru
matplotlib
numpy
omegaconf
pandas
pytorch-lightning
pre-commit
pytest
torchvision
